Struct schannel::cert_context::CertContext

pub struct CertContext(_);

Wrapper of a winapi certificate, or a PCCERT_CONTEXT.

Implementations

impl CertContext

pub fn new(data: &[u8]) -> Result<CertContext>

Decodes a DER-formatted X509 certificate.

pub fn to_der(&self) -> &[u8]

Get certificate in binary DER form

pub fn subject_public_key_info_der(&self) -> Result<Vec<u8>>

Certificate subject public key info

pub fn from_pem(pem: &str) -> Result<CertContext>

Decodes a PEM-formatted X509 certificate.

pub fn to_pem(&self) -> Result<String>

Get certificate as PEM-formatted X509 certificate.

pub fn fingerprint(&self, alg: HashAlgorithm) -> Result<Vec<u8>>

Returns a hash of this certificate

pub fn sha1(&self) -> Result<[u8; 20]>

👎 Deprecated:
please use fingerprint instead

Returns the sha1 hash of this certificate

The sha1 is returned as a 20-byte array representing the bits of the sha1 hash.

pub fn sign_hash_algorithms(&self) -> Result<String>

Returns the <SIGNATURE>/<HASH> string representing the certificate signature.

The <SIGNATURE> value identifies the CNG public key algorithm. The <HASH> value identifies the CNG hash algorithm.

Common examples are:

RSA/SHA1

RSA/SHA256

ECDSA/SHA256

pub fn signature_hash(&self) -> Result<Vec<u8>>

Returns the signature hash.

pub fn description(&self) -> Result<Vec<u8>>

Returns the property displayed by the certificate UI. This property allows the user to describe the certificate’s use.

pub fn friendly_name(&self) -> Result<String>

Returns a string that contains the display name for the certificate.

pub fn set_friendly_name(&self, name: &str) -> Result<()>

Configures the string that contains the display name for this certificate.

pub fn is_time_valid(&self) -> Result<bool>

Verifies the time validity of this certificate relative to the system’s current time.

pub fn private_key(&self) -> AcquirePrivateKeyOptions<'_>

Returns a builder used to acquire the private key corresponding to this certificate.

pub fn delete(self) -> Result<()>

Deletes this certificate from its certificate store.

pub fn set_key_prov_info(&self) -> SetKeyProvInfo<'_>

Returns a builder used to set the private key associated with this certificate.

pub fn valid_uses(&self) -> Result<ValidUses>

Returns the valid uses for this certificate

pub fn cert_store(&self) -> Option<CertStore>

For a remote certificate, returns a certificate store containing any intermediate certificates provided by the remote sender.