Struct schannel::tls_stream::Builder

pub struct Builder { /* fields omitted */ }

A builder type for TlsStreams.

Methods

impl Builder

fn new() -> Builder

Returns a new Builder.

fn domain(&mut self, domain: &str) -> &mut Builder

Sets the domain associated with connections created with this Builder.

The domain will be used for Server Name Indication as well as certificate validation.

fn cert_store(&mut self, cert_store: CertStore) -> &mut Builder

Specifies a custom certificate store which is later used when validating a server's certificate.

This option is only used for client connections and is used to construct the certificate chain which the server's certificate is validated against.

Note that adding certificates here does not mean that they are implicitly trusted. Instead certificates must sill be signed by some local trusted authority.

fn connect<S>(&mut self,
cred: SchannelCred,
stream: S)
-> Result<TlsStream<S>, HandshakeError<S>> where S: Read + Write

Initialize a new TLS session where the stream provided will be connecting to a remote TLS server.

If the stream provided is a blocking stream then the entire handshake will be performed if possible, but if the stream is in nonblocking mode then a HandshakeError::Interrupted variant may be returned. This type can then be extracted to later call MidHandshakeTlsStream::handshake when data becomes available.

fn accept<S>(&mut self,
cred: SchannelCred,
stream: S)
-> Result<TlsStream<S>, HandshakeError<S>> where S: Read + Write

Initialize a new TLS session where the stream provided will be accepting a connection.

This method will tweak the protocol for "who talks first" and also currently disables validation of the client that's connecting to us.

If the stream provided is a blocking stream then the entire handshake will be performed if possible, but if the stream is in nonblocking mode then a HandshakeError::Interrupted variant may be returned. This type can then be extracted to later call MidHandshakeTlsStream::handshake when data becomes available.