schannel

Trait RawPointer

pub trait RawPointer {
    // Required methods
    unsafe fn from_ptr(t: *mut c_void) -> Self;
    unsafe fn as_ptr(&self) -> *mut c_void;
}

Allows access to the underlying schannel API representation of a wrapped data type

Performing actions with internal handles might lead to the violation of internal assumptions and therefore is inherently unsafe.

Required Methods

unsafe fn from_ptr(t: *mut c_void) -> Self

Constructs an instance of this type from its handle / pointer.

Safety

This function is unsafe

unsafe fn as_ptr(&self) -> *mut c_void

Get a raw pointer from the underlying handle / pointer.

Safety

This function is unsafe

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.