Struct schannel::cert_context::CertContext

pub struct CertContext(_);

Wrapper of a winapi certificate, or a PCCERT_CONTEXT.

Methods

impl CertContext

fn new(data: &[u8]) -> Result<CertContext>

Creates a new certificate from the encoded form.

For example a DER-encoded certificate can be passed in to create a CertContext.

fn sha1(&self) -> Result<[u8; 20]>

Returns the sha1 hash of this certificate

The sha1 is returned as a 20-byte array representing the bits of the sha1 hash.

fn sign_hash_algorithms(&self) -> Result<String>

Returns the <SIGNATURE>/<HASH> string representing the certificate signature.

The <SIGNATURE> value identifies the CNG public key algorithm. The <HASH> value identifies the CNG hash algorithm.

Common examples are:

RSA/SHA1

RSA/SHA256

ECDSA/SHA256

fn signature_hash(&self) -> Result<Vec<u8>>

Returns the signature hash.

fn description(&self) -> Result<Vec<u8>>

Returns the property displayed by the certificate UI. This property allows the user to describe the certificate's use.

fn friendly_name(&self) -> Result<String>

Returns a string that contains the display name for the certificate.

fn set_friendly_name(&self, name: &str) -> Result<()>

Configures the string that contains the display name for this certificate.

fn is_time_valid(&self) -> Result<bool>

Verifies the time validity of this certificate relative to the system's current time.

fn private_key<'a>(&'a self) -> AcquirePrivateKeyOptions<'a>

Returns a builder used to acquire the private key corresponding to this certificate.

fn delete(self) -> Result<()>

Deletes this certificate from its certificate store.