Module schannel::cert_store

Bindings to winapi's certificate-store related APIs.

Structs

CertStore
Representation of certificate store on Windows, wrapping a HCERTSTORE.

Certs
An iterator over the certificates contained in a CertStore, returned by CertStore::iter

PfxImportOptions
A builder type for imports of PKCS #12 archives.

Enums

CertAdd
Argument to the add_cert function indicating how a certificate should be added to a CertStore.