Struct schannel::cert_store::CertStore

pub struct CertStore(_);

Representation of certificate store on Windows, wrapping a HCERTSTORE.

Methods

impl CertStore

fn open_current_user(which: &str) -> Result<CertStore>

Opens up the specified key store within the context of the current user.

Known valid values for which are "Root" and "My".

fn open_local_machine(which: &str) -> Result<CertStore>

Opens up the specified key store within the context of the local machine.

Known valid values for which are "Root" and "My".

fn import_pkcs12(data: &[u8], password: Option<&str>) -> Result<CertStore>

Imports a PKCS#12-encoded key/certificate pair, returned as a CertStore instance.

The password must also be provided to decrypt the encoded data.

fn certs(&mut self) -> Certs

Returns an iterator over the certificates in this certificate store.

fn add_cert(&mut self, cx: &CertContext, how: CertAdd) -> Result<CertContext>

Adds a certificate context to this store.

This function will add the certificate specified in cx to this store. A copy of the added certificate is returned.