Enum schannel::schannel_cred::Algorithm

#[repr(u32)]

#[non_exhaustive]
pub enum Algorithm {

    Aes,
    Aes128,
    Aes192,
    Aes256,
    AgreedkeyAny,
    CylinkMek,
    Des,
    Desx,
    DhEphem,
    DhSf,
    DssSign,
    Ecdh,
    EcdhEphem,
    Ecdsa,
    HashReplaceOwf,
    HughesMd5,
    Hmac,
    Mac,
    Md2,
    Md4,
    Md5,
    NoSign,
    Rc2,
    Rc4,
    Rc5,
    RsaKeyx,
    RsaSign,
    Sha1,
    Sha256,
    Sha384,
    Sha512,
    TripleDes,
    TripleDes112,

}

Algorithms supported by Schannel.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Aes

Advanced Encryption Standard (AES).

Aes128

128 bit AES.

Aes192

192 bit AES.

Aes256

256 bit AES.

AgreedkeyAny

Temporary algorithm identifier for handles of Diffie-Hellman–agreed keys.

CylinkMek

An algorithm to create a 40-bit DES key that has parity bits and zeroed key bits to make its key length 64 bits.

Des

DES encryption algorithm.

Desx

DESX encryption algorithm.

DhEphem

Diffie-Hellman ephemeral key exchange algorithm.

DhSf

Diffie-Hellman store and forward key exchange algorithm.

DssSign

DSA public key signature algorithm.

Ecdh

Elliptic curve Diffie-Hellman key exchange algorithm.

EcdhEphem

Ephemeral elliptic curve Diffie-Hellman key exchange algorithm.

Ecdsa

Elliptic curve digital signature algorithm.

HashReplaceOwf

One way function hashing algorithm.

HughesMd5

Hughes MD5 hashing algorithm.

Hmac

HMAC keyed hash algorithm.

Mac

MAC keyed hash algorithm.

Md2

MD2 hashing algorithm.

Md4

MD4 hashing algorithm.

Md5

MD5 hashing algorithm.

NoSign

No signature algorithm..

Rc2

RC2 block encryption algorithm.

Rc4

RC4 stream encryption algorithm.

Rc5

RC5 block encryption algorithm.

RsaKeyx

RSA public key exchange algorithm.

RsaSign

RSA public key signature algorithm.

Sha1

SHA hashing algorithm.

Sha256

256 bit SHA hashing algorithm.

Sha384

384 bit SHA hashing algorithm.

Sha512

512 bit SHA hashing algorithm.

TripleDes

Triple DES encryption algorithm.

TripleDes112

Two-key triple DES encryption with effective key length equal to 112 bits.